+++ tests/e2e/test_15_http.py
@@ -2,22 +2,12 @@ from typing import Any, Generator
 
 import pytest
 
+from testkit.helpers import hostname_prefix, stop_all_runs, wait_until_server_ready
+from testkit.suites import TestRPC_E2E as BaseRPC_E2E
+from testkit.things import TestThing
+
 from hololinked.client import ClientFactory, ObjectProxy
 from hololinked.utils import uuid_hex
-
-
-try:
-    from .conftest import stop_all_runs
-    from .test_11_rpc_e2e import TestRPC_E2E as BaseRPC_E2E  # noqa: F401
-    from .test_11_rpc_e2e import client, thing, thing_model  # noqa: F401
-    from .test_14_protocols_http import hostname_prefix, wait_until_server_ready
-    from .things import TestThing
-except ImportError:
-    from conftest import stop_all_runs
-    from test_11_rpc_e2e import TestRPC_E2E as BaseRPC_E2E  # noqa: F401
-    from test_11_rpc_e2e import client, thing, thing_model  # noqa: F401
-    from test_14_protocols_http import hostname_prefix, wait_until_server_ready
-    from things import TestThing
 
 
 @pytest.fixture(scope="class")
