+++ tests/e2e/test_16_mqtt.py
@@ -1,6 +1,4 @@
 import itertools
-import os
-import ssl
 import time
 
 from typing import Any, Generator
@@ -10,6 +8,13 @@ import pytest
 from testcontainers.mqtt import (
     MosquittoContainer,  # TODO this will not work from the current release of testcontainers
 )
+from testkit.helpers import (  # noqa: F401
+    hostname_prefix,
+    mqtt_ssl_context,
+    stop_all_runs,
+    wait_until_server_ready,
+)
+from testkit.things import TestThing
 
 from hololinked.client import ClientFactory, ObjectProxy
 from hololinked.client.abstractions import SSE
@@ -19,90 +24,10 @@ from hololinked.server.mqtt.server import MQTTPublisher
 from hololinked.utils import uuid_hex
 
 
-try:
-    from tests.conftest import stop_all_runs
-    from tests.test_14_protocols_http import (  # noqa: F401
-        hostname_prefix,
-        wait_until_server_ready,
-    )
-    from tests.things import TestThing
-except ImportError:
-    from conftest import stop_all_runs
-    from test_14_protocols_http import (  # noqa: F401
-        hostname_prefix,
-        wait_until_server_ready,
-    )
-    from things import TestThing
+pytestmark = pytest.mark.mqtt
 
 
 count = itertools.count(63500)
-
-
-def mqtt_ssl_context() -> ssl.SSLContext:
-    mqtt_ssl = ssl.create_default_context(ssl.Purpose.SERVER_AUTH)
-    if not os.path.exists(f"daq-system-infrastructure{os.sep}certs{os.sep}ca.crt"):
-        raise FileNotFoundError("CA certificate 'ca.crt' not found in current directory for MQTT TLS connection")
-    mqtt_ssl.load_verify_locations(cafile=f"daq-system-infrastructure{os.sep}certs{os.sep}ca.crt")
-    mqtt_ssl.verify_mode = ssl.CERT_REQUIRED
-    mqtt_ssl.minimum_version = ssl.TLSVersion.TLSv1_2
-    return mqtt_ssl
-
-
-@pytest.fixture(scope="module")
-def mosquitto_container() -> Generator[MosquittoContainer, None, None]:
-    container = MosquittoContainer(
-        volumes=[
-            (
-                os.path.abspath("daq-system-infrastructure/conf/mosquitto.conf"),
-                "/mosquitto/config/mosquitto.conf",
-                "ro",
-            ),
-            (os.path.abspath("daq-system-infrastructure/conf/passwords.txt"), "/mosquitto/config/passwords.txt", "ro"),
-            (os.path.abspath("daq-system-infrastructure/data/mosquitto"), "/mosquitto/data", "rw"),
-            (os.path.abspath("daq-system-infrastructure/data/mosquitto/log"), "/mosquitto/log", "rw"),
-            (os.path.abspath("daq-system-infrastructure/data/mosquitto/persisted"), "/mosquitto/data/persisted", "rw"),
-            (os.path.abspath("daq-system-infrastructure/certs"), "/mosquitto/config/certs", "ro"),
-        ],
-        username="sampleuser",
-        password="samplepass",
-        mqtt_port=8883,
-        ssl_context=mqtt_ssl_context(),
-    )
-
-    # One could in principle retry the tests directly
-    for i in range(3):
-        exc = None
-        try:
-            container.start()
-            break
-        except Exception as ex:
-            # some SSL handshake error appears sometimes, but not always
-            print(f"\nAttempt {i + 1}: Failed to start Mosquitto container, retrying...")
-            exc = ex
-        try:
-            print(container.get_logs())
-        except Exception:
-            print("Failed to get container logs")
-        try:
-            container.stop()
-        except Exception:
-            pass
-        if i == 2:
-            raise exc from None
-        time.sleep(3)
-
-    yield container
-    container.stop()
-
-
-@pytest.fixture(scope="module")
-def mqtt_host(mosquitto_container: MosquittoContainer) -> str:
-    return mosquitto_container.get_container_host_ip()
-
-
-@pytest.fixture(scope="module")
-def mqtt_port(mosquitto_container: MosquittoContainer) -> int:
-    return int(mosquitto_container.get_exposed_port(8883))
 
 
 @pytest.fixture(scope="module")
