+++ tests/testkit/helpers/misc.py
