+++ tests/testkit/suites/__init__.py
@@ -1,0 +1,1 @@
+from .rpc import TestRPC_E2E  # noqa: F401
