+++ tests/testkit/things/spectrometer.py
