+++ tests/unit/server/zmq/test_01_message.py
@@ -1,6 +1,13 @@
-from uuid import UUID, uuid4
+from uuid import uuid4
 
 import pytest
+
+from testkit.helpers import (
+    AppIDs,
+    validate_event_message,
+    validate_request_message,
+    validate_response_message,
+)
 
 from hololinked import Serializers
 from hololinked.server.zmq.message import (
@@ -11,87 +18,12 @@ from hololinked.server.zmq.message import (
     OPERATION,
     REPLY,
     TIMEOUT,
-    EventHeader,
     EventMessage,
     PreserializedData,
-    RequestHeader,
     RequestMessage,
-    ResponseHeader,
     ResponseMessage,
     SerializableData,
 )
-
-
-try:
-    from .conftest import AppIDs
-except ImportError:
-    from conftest import AppIDs
-
-
-def validate_request_message(request_message: RequestMessage, app_ids: AppIDs) -> None:
-    """Call this method to validate request message"""
-    # req. 1. check message ID is a UUID
-    assert isinstance(request_message.id, UUID) or isinstance(UUID(request_message.id, version=4), UUID)
-    # req. 2. generated byte array must confine to predefined length (which is readonly & fixed)
-    assert len(request_message.byte_array) == request_message.length
-    # req. 3. receiver which must be the server ID
-    assert request_message.receiver_id == app_ids.server_id
-    # req. 4. sender_id is the client ID
-    assert request_message.sender_id == app_ids.client_id
-    # req. 5. all indices of byte array are bytes
-    for obj in request_message.byte_array:
-        assert isinstance(obj, bytes)
-    # req. 6. check that header is correct type (RequestHeader dataclass/struct)
-    assert isinstance(request_message.header, RequestHeader)
-    # req. 7 check that body is correct type (list of SerializableData and PreserializedData)
-    assert isinstance(request_message.body, list)
-    assert len(request_message.body) == 2
-    assert isinstance(request_message.body[0], SerializableData)
-    assert isinstance(request_message.body[1], PreserializedData)
-
-
-def validate_response_message(response_message: ResponseMessage, app_ids: AppIDs) -> None:
-    """Call this method to validate response message"""
-    # check message ID is a UUID
-    assert isinstance(response_message.id, UUID) or isinstance(UUID(response_message.id, version=4), UUID)
-    # check message length
-    assert len(response_message.byte_array) == response_message.length
-    # check receiver which must be the client
-    assert response_message.receiver_id == app_ids.client_id
-    # sender_id is not set before sending message on the socket
-    assert response_message.sender_id == app_ids.server_id
-    # check that all indices are bytes
-    for obj in response_message.byte_array:
-        assert isinstance(obj, bytes)
-    # check that header is correct type
-    assert isinstance(response_message.header, ResponseHeader)
-    # check that body is correct type
-    assert isinstance(response_message.body, list)
-    assert len(response_message.body) == 2
-    assert isinstance(response_message.body[0], SerializableData)
-    assert isinstance(response_message.body[1], PreserializedData)
-
-
-def validate_event_message(event_message: EventMessage, app_ids: AppIDs) -> None:
-    """Call this method to validate event message"""
-    # check message ID is a UUID
-    assert isinstance(event_message.id, UUID) or isinstance(UUID(event_message.id, version=4), UUID)
-    # check message length
-    assert len(event_message.byte_array) == event_message.length
-    # no receiver id for event message, only event id
-    assert isinstance(event_message.event_id, str)
-    # sender_id is not set before sending message on the socket
-    assert event_message.sender_id == app_ids.server_id
-    # check that all indices are bytes
-    for obj in event_message.byte_array:
-        assert isinstance(obj, bytes)
-    # check that header is correct type
-    assert isinstance(event_message.header, EventHeader)
-    # check that body is correct type
-    assert isinstance(event_message.body, list)
-    assert len(event_message.body) == 2
-    assert isinstance(event_message.body[0], SerializableData)
-    assert isinstance(event_message.body[1], PreserializedData)
 
 
 def test_01_request_message(app_ids: AppIDs) -> None:
