+++ tests/unit/server/zmq/test_02_socket.py
